requirements.txt
discord.py>=2.3.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
gspread>=6.0.0
google-auth>=2.0.0
